Set unused fingerprint entries to random values
This slightly reduces the false-positive rate of entry lookups compared
to when unused entries are set to 0. For a Fuse8 filter, the
false-positive rate drops from ~0.39% to ~0.37%.

The commit also exposes an analysis module for measuring the
distribution of zeroes in a filter, thanks to @rot256's insights in #16.

Closes #16

cc @rot256