Remove inline annotations from larger functions
I think I was a bit too eager with hinting inlines here, and would like
to do a bit more analysis before deciding if this is the right way to
go. Right now I'm not convinced that it is; it feels like trying to
outsmart rustc, which is probably a bad idea.